Refactor Streaming Calls to manager classes
