+++ android/sdl_android/src/main/java/com/smartdevicelink/SdlConnection/SdlSession.java
@@ -35,21 +35,11 @@ package com.smartdevicelink.SdlConnection;
 import android.content.Context;
 
 import com.smartdevicelink.exception.SdlException;
-import com.smartdevicelink.protocol.ProtocolMessage;
 import com.smartdevicelink.protocol.SdlPacket;
 import com.smartdevicelink.protocol.SdlProtocol;
 import com.smartdevicelink.protocol.SdlProtocolBase;
 import com.smartdevicelink.protocol.enums.SessionType;
-import com.smartdevicelink.proxy.interfaces.IAudioStreamListener;
 import com.smartdevicelink.proxy.interfaces.ISdlServiceListener;
-import com.smartdevicelink.proxy.interfaces.IVideoStreamListener;
-import com.smartdevicelink.proxy.rpc.VideoStreamingFormat;
-import com.smartdevicelink.proxy.rpc.enums.VideoStreamingProtocol;
-import com.smartdevicelink.streaming.AbstractPacketizer;
-import com.smartdevicelink.streaming.IStreamListener;
-import com.smartdevicelink.streaming.StreamPacketizer;
-import com.smartdevicelink.streaming.video.RTPH264Packetizer;
-import com.smartdevicelink.streaming.video.VideoStreamingParameters;
 import com.smartdevicelink.transport.MultiplexTransportConfig;
 import com.smartdevicelink.transport.TCPTransportConfig;
 import com.smartdevicelink.transport.enums.TransportType;
@@ -57,7 +47,6 @@ import com.smartdevicelink.util.DebugTool;
 import com.smartdevicelink.util.MediaStreamingStatus;
 import com.smartdevicelink.util.Version;
 
-import java.io.IOException;
 import java.lang.ref.WeakReference;
 import java.util.concurrent.CopyOnWriteArrayList;
 
@@ -178,13 +167,8 @@ public class SdlSession extends BaseSdlSession {
 
     @Override
     public void onServiceEnded(SdlPacket packet, SessionType serviceType, int sessionID) {
-
         if (SessionType.RPC.equals(serviceType)) {
             this.sessionListener.onSessionEnded(sessionID);
-        } else if (SessionType.NAV.equals(serviceType)) {
-            stopVideoStream();
-        } else if (SessionType.PCM.equals(serviceType)) {
-            stopAudioStream();
         }
 
         if (serviceListeners != null && serviceListeners.containsKey(serviceType)) {
@@ -198,12 +182,6 @@ public class SdlSession extends BaseSdlSession {
 
     @Override
     public void onServiceError(SdlPacket packet, SessionType serviceType, int sessionID, String error) {
-        if (SessionType.NAV.equals(serviceType)) {
-            stopVideoStream();
-        } else if (SessionType.PCM.equals(serviceType)) {
-            stopAudioStream();
-        }
-
         if (serviceListeners != null && serviceListeners.containsKey(serviceType)) {
             CopyOnWriteArrayList<ISdlServiceListener> listeners = serviceListeners.get(serviceType);
             for (ISdlServiceListener listener : listeners) {
@@ -214,87 +192,5 @@ public class SdlSession extends BaseSdlSession {
 
     @Override
     public void onAuthTokenReceived(String authToken) {/* Do nothing */ }
-
-    /* ***********************************************************************************************************************************************************************
-     * *****************************************************************  Fix after initial refactor *********************************************************************************
-     *************************************************************************************************************************************************************************/
-    //FIXME there is a lot of spaghetti code here that needs to be addressed. For first refactor the
-    // the goal is to only refactor SdlSession. Another PR should be opened to fix all the packetizer
-    // classes and method calls.
-
-    //FIXME Move this logic to the related streaming manager
-    private AbstractPacketizer videoPacketizer;
-    private StreamPacketizer audioPacketizer;
-
-    IStreamListener streamListener = new IStreamListener() {
-        @Override
-        public void sendStreamPacket(ProtocolMessage pm) {
-            sendMessage(pm);
-        }
-    };
-
-    private VideoStreamingProtocol getAcceptedProtocol() {
-        if (acceptedVideoParams != null) {
-            VideoStreamingFormat format = acceptedVideoParams.getFormat();
-            if (format != null && format.getProtocol() != null) {
-                return format.getProtocol();
-            }
-        }
-        //Returns default protocol if none are found
-        return new VideoStreamingParameters().getFormat().getProtocol();
-
-    }
-
-    public IVideoStreamListener startVideoStream() {
-        VideoStreamingProtocol protocol = getAcceptedProtocol();
-        try {
-            switch (protocol) {
-                case RAW: {
-                    videoPacketizer = new StreamPacketizer(streamListener, null, SessionType.NAV, (byte) this.sessionId, this);
-                    videoPacketizer.start();
-                    return (IVideoStreamListener) videoPacketizer;
-                }
-                case RTP: {
-                    //FIXME why is this not an extension of StreamPacketizer?
-                    videoPacketizer = new RTPH264Packetizer(streamListener, SessionType.NAV, (byte) this.sessionId, this);
-                    videoPacketizer.start();
-                    return (IVideoStreamListener) videoPacketizer;
-                }
-                default:
-                    DebugTool.logError(TAG, "Protocol " + protocol + " is not supported.");
-                    return null;
-            }
-        } catch (IOException e) {
-            return null;
-        }
-    }
-
-    public IAudioStreamListener startAudioStream() {
-        try {
-            audioPacketizer = new StreamPacketizer(streamListener, null, SessionType.PCM, (byte) this.sessionId, this);
-            audioPacketizer.start();
-            return audioPacketizer;
-        } catch (IOException e) {
-            return null;
-        }
-
-    }
-
-
-    public boolean stopVideoStream() {
-        if (videoPacketizer != null) {
-            videoPacketizer.stop();
-            return true;
-        }
-        return false;
-    }
-
-    public boolean stopAudioStream() {
-        if (audioPacketizer != null) {
-            audioPacketizer.stop();
-            return true;
-        }
-        return false;
-    }
 
 }
